refactored onAptUpdate and appointments into context for accessibility across pages
