+++ utils/context/appointmentsContext.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const AppointmentsContext = createContext();
+
+export default AppointmentsContext;
